moved generators to domain
