+++ src/Commands/Generators/AbstractGeneratorCommand.php
@@ -1,7 +1,7 @@
 <?php
 // Created by dealloc. All rights reserved.
 
-namespace Experus\Sockets\Commands;
+namespace Experus\Sockets\Commands\Generators;
 
 use Illuminate\Console\Command;
 use Illuminate\Contracts\Foundation\Application;
