+++ src/Commands/Generators/GenerateStackCommand.php
@@ -1,7 +1,7 @@
 <?php
 // Created by dealloc. All rights reserved.
 
-namespace Experus\Sockets\Commands;
+namespace Experus\Sockets\Commands\Generators;
 
 
 class GenerateStackCommand extends AbstractGeneratorCommand
